userpublicprofileresponse 삭제
userprofileresponse 재사용, 모든 유저 -> 아티스트만 거르게 수정
